Refactor ExprBuilder into its own module
